Implement negative toggle, and delete
fixed code to be able to maintain input and number values so that
multiple decimals could not be added.

needed to adjust the index of nested loop on compute function to account
for adjusted "input" values. once an operation is performed, you have to
loop through the input array starting where you left off, not the next
value